code: updated all workloads to use int
This change updates all the algorithms to use int. In addition, the
code is tested with large (~178MiB) graphs.

+++ src/bfs.hh
@@ -9,7 +9,7 @@ namespace simple {
 class BFS  {
     private:
     public:
-        BFS(Graph *G, uint64_t starting_node);
+        BFS(Graph *G, int starting_node);
 };
 }
 #endif
